backend/pyproject.toml
[tool.poetry]
name = "autobrowse-backend"
version = "0.1.0"
description = "FastAPI Backend for AutoBrowse"
authors = ["AutoBrowse Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.1"
pydantic-settings = "^2.2.1"
supabase = "^2.3.4"
python-socketio = "^5.11.1"
python-dotenv = "^1.0.1"
browser-use = {version = "^0.1.40", extras = ["memory"]}
langchain-anthropic = "^0.3.3"
langchain-openai = "^0.3.0"
playwright = "^1.42.0"
pyjwt = "^2.10.1"
python-multipart = "^0.0.21"
cryptography = "^46.0.3"
docker = "^7.1.0"
aiohttp = "^3.13.3"
opencv-python-headless = "^4.13.0.90"
ultralytics = "^8.4.6"
duckduckgo-search = "^8.1.1"
beautifulsoup4 = "^4.14.3"
tenacity = "^9.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
